Inline compile_parameters back into Shrine compiler
