+++ lib/tapioca/dsl/compilers/shrine.rb
@@ -3,8 +3,6 @@
 
 return unless defined?(Shrine)
 
-require "tapioca/dsl/helpers/parameter_compilation"
-
 module Tapioca
   module Dsl
     module Compilers
@@ -49,8 +47,6 @@ module Tapioca
       # end
       # ~~~
       class Shrine < Tapioca::Dsl::Compiler
-        include Helpers::ParameterCompilation
-
         InstanceMethodModuleName = "ShrineGeneratedMethods"
         ClassMethodModuleName = "ShrineGeneratedClassMethods"
 
@@ -148,6 +144,35 @@ module Tapioca
             "T::Boolean"
           else
             "T.untyped"
+          end
+        end
+
+        # Compiles a `Method` / `UnboundMethod#parameters` into the `RBI::TypedParam`
+        # array expected by `RBI::Scope#create_method`. Argument types default to
+        # `T.untyped` since shrine plugins don't carry sigs. Anonymous parameter names
+        # — including empty names and the special tokens `:*`, `:**`, `:&` — are
+        # replaced with `_arg{index}` so the generated RBI is syntactically valid.
+        #: ((Method | UnboundMethod) method_obj) -> Array[RBI::TypedParam]
+        def compile_parameters(method_obj)
+          method_obj.parameters.each_with_index.filter_map do |(type, name), index|
+            name_str = name.to_s
+            name_str = "_arg#{index}" unless name_str.match?(/\A[A-Za-z_][A-Za-z0-9_]*\z/)
+            case type
+            when :req
+              create_param(name_str, type: "T.untyped")
+            when :opt
+              create_opt_param(name_str, type: "T.untyped", default: "T.unsafe(nil)")
+            when :rest
+              create_rest_param(name_str, type: "T.untyped")
+            when :keyreq
+              create_kw_param(name_str, type: "T.untyped")
+            when :key
+              create_kw_opt_param(name_str, type: "T.untyped", default: "T.unsafe(nil)")
+            when :keyrest
+              create_kw_rest_param(name_str, type: "T.untyped")
+            when :block
+              create_block_param(name_str, type: "T.untyped")
+            end
           end
         end
       end
